# remote_cache/digest: add benchmark for sha256-simd