Provide a setup to compare minio/sha256-simd (Apache 2.0 license)
performance vs the Go standard library "crypto/sha256".

The `sha256-simd` library comes with 2 modes:
- without server, automatically detect CPU features
- with server, require Avx512 CPU features

The ARM64 support is not tested.

Running the benchmark against out remote executor yields

```
==================== Test output for //server/remote_cache/digest:simd_bench_test:
goos: linux
goarch: amd64
cpu: Intel(R) Xeon(R) CPU @ 3.10GHz

BenchmarkSIMDDigestCompute/without_SIMD/1-30                      255240              5042 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_no_server/1-30               234526              5190 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_with_server/1-30                388          36804140 ns/op

BenchmarkSIMDDigestCompute/without_SIMD/10-30                      10000            118668 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_no_server/10-30               26204             64872 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_with_server/10-30               100          62445228 ns/op

BenchmarkSIMDDigestCompute/without_SIMD/100-30                     10000            193471 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_no_server/100-30              20247            135334 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_with_server/100-30              100          64685802 ns/op

BenchmarkSIMDDigestCompute/without_SIMD/1000-30                    14314            188163 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_no_server/1000-30             10000            176901 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_with_server/1000-30             100         212289431 ns/op

BenchmarkSIMDDigestCompute/without_SIMD/10000-30                    9067            658089 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_no_server/10000-30            10000            721403 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_with_server/10000-30            100         234613900 ns/op

BenchmarkSIMDDigestCompute/without_SIMD/100000-30                   2685           1577976 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_no_server/100000-30            1924           1079974 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_with_server/100000-30           100         146595705 ns/op

BenchmarkSIMDDigestCompute/without_SIMD/1000000-30                   312           9117083 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_no_server/1000000-30            298          13086220 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_with_server/1000000-30           56         211401036 ns/op

PASS
================================================================================
```

<!-- Optional: Provide additional context (beyond the PR title). -->

<!-- Optional: link a GitHub issue.
     Example: "Fixes #123" will auto-close #123 when the PR is merged. -->

**Related issues**: N/A