@sluongng 2023-08-14
```
==================== Test output for //server/remote_cache/digest:simd_bench_test:
goos: linux
goarch: amd64
cpu: Intel(R) Xeon(R) CPU @ 3.10GHz
BenchmarkSIMDDigestCompute/without_SIMD/1000000-30                          2180            556758 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_no_server/1000000-30                   2282            674404 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_with_server/1000000-30                  571           2252739 ns/op

BenchmarkSIMDDigestCompute/without_SIMD/10000000-30                          121           8813178 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_no_server/10000000-30                   129           8296944 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_with_server/10000000-30                  50          20486731 ns/op

BenchmarkSIMDDigestCompute/without_SIMD/100000000-30                          14          78125744 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_no_server/100000000-30                   13          80598781 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_with_server/100000000-30                  7         177869732 ns/op

BenchmarkSIMDDigestCompute/without_SIMD/1000000000-30                          1        4463899572 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_no_server/1000000000-30                   1        1478056796 ns/op
BenchmarkSIMDDigestCompute/with_SIMD_with_server/1000000000-30                 1        1513101099 ns/op
================================================================================
```

Since the doc mentioned speed up for cases `>1MB`, I tried to run the test against some larger loads.

Overall, the constraint of (1) a bigger size message, (2) server - CPU core 1-1 mapping, and (3) message padding for alignment make it quite unattractive. to our use case.

Gona close this for now.