@sluongng 2023-08-16
After digging into this a bit more, it seems like the CPU we have on GCP, at least for our executor, do not include Intel's SHA extension

```
Name: Intel(R) Xeon(R) CPU @ 3.10GHz
PhysicalCores: 15
ThreadsPerCore: 2
LogicalCores: 30
Family 6 Model: 85 Vendor ID: Intel
Features: ADX,AESNI,AVX,AVX2,AVX512BW,AVX512CD,AVX512DQ,AVX512F,AVX512VL,AVX512VNNI,BMI1,BMI2,CLMUL,CMOV,CMPXCHG8,CX16,ERMS,F16C,FMA3,FXSR,FXSROPT,HLE,HTT,HYPERVISOR,IA32_ARCH_CAP,IBPB,LAHF,LZCNT,MD_CLEAR,MMX,MOVBE,MPX,NX,OSXSAVE,POPCNT,RDRAND,RDSEED,RDTSCP,RTM,SPEC_CTRL_SSBD,SSE,SSE2,SSE3,SSE4,SSE42,SSSE3,STIBP,SYSCALL,SYSEE,VMX,X87,XGETBV1,XSAVE,XSAVEC,XSAVEOPT,XSAVES
Cacheline bytes: 64
L1 Data Cache: 32768 bytes
L1 Instruction Cache: 32768 bytes
L2 Cache: 1048576 bytes
L3 Cache: 25952256 bytes
Frequency 3100000000 hz
```

And the minio/sha256-simd code has this clause https://github.com/minio/sha256-simd/blob/6096f891a77bfe490cbea7a424c821b5fdb92849/cpuid_other.go#L27

So when we use `sha256simd.New()`, that is essentially a thin wrap around `crypto/sha256`, and thus, the result made no difference. If we ever made a switch to AMD Ryzen / Epyc, we could test this again.

The AVX512 implementation is mostly targeted toward hashing bigger files/messages and thus is not suitable for our use case for now.

The ARM64 implementation could be attractive for ARM64 executors (Linux / MacOS) down the line, but my benchmark on M1 laptop does not show a big speed-up.

---

Pushed my latest local setup to the branch so future me / other folks could replicate the experiment.